requirements.txt
torch
torchvision
math
Pillow
os
matplotlib		
numpy	
opencv-python			
scikit-learn	